KAFKA-2896: moved parse_describe_topic into KafkaService and using one security protocol for partition re-assignment test
